install lastest release

+++ linux_install.sh
@@ -1,8 +1,9 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 set -e
 
-VERSION="1.0.0"
+REPO="KiamMota/lpm"
+
 OS=$(uname -s | tr '[:upper:]' '[:lower:]')
 ARCH=$(uname -m)
 
@@ -21,24 +22,39 @@ if [ "$OS" != "linux" ] && [ "$OS" != "darwin" ]; then
   exit 1
 fi
 
-URL="https://github.com/KiamMota/lpm/releases/download/v${VERSION}/lpm_${VERSION}_${OS}_${ARCH}.tar.gz"
+echo "==> detecting latest release..."
 
-echo "==> downloading lpm ${VERSION} (${OS}/${ARCH})..."
-curl -fsSL "$URL" -o /tmp/lpm.tar.gz
+VERSION=$(curl -fsSL https://api.github.com/repos/$REPO/releases/latest | grep '"tag_name"' | sed -E 's/.*"v([^"]+)".*/\1/')
 
-echo "==> extracting..."
-tar -xzf /tmp/lpm.tar.gz -C /tmp
+if [ -n "$VERSION" ]; then
+  echo "==> latest version: $VERSION"
 
-if command -v lpm &>/dev/null; then
-  echo "==> removing old lpm..."
-  sudo rm -f "$(which lpm)"
+  URL="https://github.com/$REPO/releases/download/v${VERSION}/lpm_${VERSION}_${OS}_${ARCH}.tar.gz"
+
+  echo "==> downloading..."
+  curl -fsSL "$URL" -o /tmp/lpm.tar.gz
+
+  echo "==> extracting..."
+  tar -xzf /tmp/lpm.tar.gz -C /tmp
+
+  echo "==> installing..."
+  sudo mv /tmp/lpm /usr/local/bin/lpm
+  sudo chmod +x /usr/local/bin/lpm
+
+  rm -f /tmp/lpm.tar.gz
+
+else
+  echo "==> failed to get release, compiling with go..."
+
+  if ! command -v go &>/dev/null; then
+    echo "error: go not installed"
+    exit 1
+  fi
+
+  go install github.com/$REPO@latest
+
+  sudo mv "$(go env GOPATH)/bin/lpm" /usr/local/bin/lpm
 fi
 
-echo "==> installing..."
-sudo mv /tmp/lpm /usr/local/bin/lpm
-sudo chmod +x /usr/local/bin/lpm
-
-rm -f /tmp/lpm.tar.gz
-
-echo "==> lpm ${VERSION} installed!"
+echo "==> installed:"
 lpm --version
